added basic menu system
