+++ ScreensEnum.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+enum ScreensEnum
+{
+	Menu, LevelSelection, Play, HighScores, Credits, None
+};
+
