+++ Tetris.vcxproj
@@ -173,12 +173,14 @@
     <ClInclude Include="Grid.hpp" />
     <ClInclude Include="HighScoreTable.hpp" />
     <ClInclude Include="HighScoreTableScreen.hpp" />
+    <ClInclude Include="Input.hpp" />
     <ClInclude Include="Level.hpp" />
     <ClInclude Include="LevelSelectionScreen.hpp" />
     <ClInclude Include="MenuScreen.hpp" />
     <ClInclude Include="GameScreen.hpp" />
     <ClInclude Include="Score.hpp" />
     <ClInclude Include="Screen.hpp" />
+    <ClInclude Include="ScreensEnum.hpp" />
     <ClInclude Include="Sound.hpp" />
     <ClInclude Include="Tetromino.hpp" />
     <ClInclude Include="TextElement.hpp" />
